Fix shared-malloc with huge pages
This fixes a bug introduced in b024b9e443552c94609fc56a240cdb199a329853

The constant HUGE_PAGE_SIZE needs to be 64 bits, otherwise the following
has an integer overflow (line 187 in smpi_shared.cpp):
    mem = (void*)ALIGN_UP(allocated_ptr, HUGE_PAGE_SIZE);

This commit also modifies the declaration of PAGE_SIZE for the sake of
uniformity.

+++ src/smpi/internals/smpi_shared.cpp
@@ -145,8 +145,8 @@ static void *smpi_shared_malloc_local(size_t size, const char *file, int line)
 #define ALIGN_UP(n, align) (((int64_t)(n) + (int64_t)(align) - 1) & -(int64_t)(align))
 #define ALIGN_DOWN(n, align) ((int64_t)(n) & -(int64_t)(align))
 
-constexpr unsigned PAGE_SIZE      = 0x1000;
-constexpr unsigned HUGE_PAGE_SIZE = 1U << 21;
+constexpr long unsigned PAGE_SIZE      = 0x1000;
+constexpr long unsigned HUGE_PAGE_SIZE = 1U << 21;
 
 /* Similar to smpi_shared_malloc, but only sharing the blocks described by shared_block_offsets.
  * This array contains the offsets (in bytes) of the block to share.
